+++ src/components/_classes/component/Component.js
@@ -932,12 +932,16 @@ export default class Component extends Element {
    * @param string
    * @returns {*}
    */
-  sanitize(dirty, forceSanitize) {
+  sanitize(dirty, forceSanitize, options) {
     // No need to sanitize when generating PDF'S since no users interact with the form.
     if ((!this.shouldSanitizeValue && !forceSanitize) || ((this.options.pdf) && !forceSanitize)) {
       return dirty;
     }
-    return FormioUtils.sanitize(dirty, this.options);
+    return FormioUtils.sanitize(
+      dirty,
+      {
+        sanitizeConfig: _.merge(this.options?.sanitizeConfig || {}, options || {}),
+      });
   }
 
   /**
@@ -1107,13 +1111,9 @@ export default class Component extends Element {
         this.tooltips[index] = tippy(tooltip, {
           trigger: 'mouseenter click focus',
           placement: 'right',
-          allowHTML: true,
-          arrow: true,
-          content: `
-            <div class="tooltip" style="opacity: 1;" role="tooltip">
-              <div class="tooltip-arrow"></div>
-              <div class="tooltip-inner">${this.t(tooltipText, { _userInput: true })}</div>
-            </div>`,
+          zIndex: 10000,
+          interactive: true,
+          content: this.t(tooltipText, { _userInput: true }),
         });
       }
     });
@@ -1616,9 +1616,9 @@ export default class Component extends Element {
     return (this.component.errors && this.component.errors[type]) ? this.component.errors[type] :  type;
   }
 
-  setContent(element, content, forceSanitize) {
+  setContent(element, content, forceSanitize, sanitizeOptions) {
     if (element instanceof HTMLElement) {
-      element.innerHTML = this.sanitize(content, forceSanitize);
+      element.innerHTML = this.sanitize(content, forceSanitize, sanitizeOptions);
       return true;
     }
     return false;
@@ -2858,7 +2858,7 @@ export default class Component extends Element {
 
   setComponentValidity(messages, dirty, silentCheck) {
     const hasErrors = !!messages.filter(message => message.level === 'error' && !message.fromServer).length;
-    if (messages.length && (!silentCheck || this.error) && (dirty || !this.pristine)) {
+    if (messages.length && (!silentCheck || this.error) && (!this.isEmpty(this.defaultValue) || dirty || !this.pristine)) {
       this.setCustomValidity(messages, dirty);
     }
     else if (!silentCheck) {