+++ src/widgets/CalendarWidget.js
@@ -357,7 +357,11 @@ export default class CalendarWidget extends InputWidget {
       return 'Invalid Date';
     }
 
-    return this.getDateValue(dates[0], this.valueFormat);
+    return (this.settings.saveAs === 'text') ? this.getDateValue(dates[0], this.valueFormat) : dates[0];
+  }
+
+  isValueISO8601(value) {
+    return value && (typeof value === 'string') && value.match(/-[0-9]{2}T[0-9]{2}:/);
   }
 
   /**
@@ -366,16 +370,17 @@ export default class CalendarWidget extends InputWidget {
    * @param value
    */
   setValue(value) {
+    const saveAsText = (this.settings.saveAs === 'text');
     if (!this.calendar) {
       value = value ? formatDate(value, convertFormatToMoment(this.settings.format), this.timezone, convertFormatToMoment(this.valueMomentFormat)) : value;
       return super.setValue(value);
     }
-
-    if (value && typeof value === 'string' && !this.settings.enableTime && value.includes('T')) {
-      this.calendar.setDate(value.split('T')[0], false);
+    if (this.isValueISO8601(value)) {
+      this.calendar.setDate(moment(value).toDate(), false);
     }
     else if (value) {
-      if ((this.settings.saveAs !== 'text') && this.settings.readOnly && !this.loadZones()) {
+      const zonesLoading = this.loadZones();
+      if (!saveAsText && this.settings.readOnly && !zonesLoading) {
         this.calendar.setDate(momentDate(value, this.valueFormat, this.timezone).toDate(), false);
       }
       else {
@@ -390,7 +395,7 @@ export default class CalendarWidget extends InputWidget {
   getValueAsString(value, format) {
     const inputFormat = format || this.dateFormat;
     const valueFormat = this.calendar ? this.valueFormat : this.settings.dateFormat;
-    if (this.settings.saveAs === 'text') {
+    if (this.settings.saveAs === 'text' && this.componentInstance.parent && !this.settings.readOnly) {
       return moment(value, convertFormatToMoment(valueFormat)).format(convertFormatToMoment(valueFormat));
     }
     return formatDate(value, inputFormat, this.timezone, convertFormatToMoment(valueFormat));
@@ -436,8 +441,7 @@ export default class CalendarWidget extends InputWidget {
     this.calendar = new Flatpickr(this._input, { ...this.settings, disableMobile: true });
 
     if (dateValue) {
-      const valueFormat = (this.settings.saveAs === 'text') ? (this.calendar ? this.valueFormat : this.settings.dateFormat) : this.dateFormat;
-      this.calendar.setDate(moment(dateValue, convertFormatToMoment(valueFormat)).toDate(), false, this.settings.altFormat);
+      this.calendar.setDate(moment(dateValue, convertFormatToMoment(this.dateFormat)).toDate(), false, this.settings.altFormat);
     }
 
     this.calendar.altInput.addEventListener('input', (event) => {
@@ -531,7 +535,7 @@ export default class CalendarWidget extends InputWidget {
     return (date, format) => {
       // Only format this if this is the altFormat and the form is readOnly.
       if (this.settings.readOnly && (format === this.settings.altFormat)) {
-        if (this.settings.saveAs === 'text' || !this.settings.enableTime || this.loadZones()) {
+        if (!this.settings.enableTime || this.loadZones()) {
           return Flatpickr.formatDate(date, format);
         }